requirements.txt
streamlit>=1.34.0
pandas>=2.2.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.1
isodate>=0.6.1
langchain-google-genai>=0.0.6
